15: Extract asset fixtures
